Copy padding to looprestoration.rs
