+++ src/looprestoration_tmpl_16.rs
@@ -134,10 +134,6 @@ use crate::src::tables::dav1d_sgr_x_by_x;
 pub type pixel = uint16_t;
 pub type coef = int32_t;
 use crate::src::looprestoration::LrEdgeFlags;
-use crate::src::looprestoration::LR_HAVE_BOTTOM;
-use crate::src::looprestoration::LR_HAVE_LEFT;
-use crate::src::looprestoration::LR_HAVE_RIGHT;
-use crate::src::looprestoration::LR_HAVE_TOP;
 pub type const_left_pixel_row = *const [pixel; 4];
 use crate::src::looprestoration::LooprestorationParams;
 
@@ -146,6 +142,8 @@ use crate::include::common::intops::iclip;
 use crate::include::common::intops::imax;
 use crate::include::common::intops::umin;
 use crate::src::looprestoration::Dav1dLoopRestorationDSPContext;
+use crate::src::looprestoration::padding;
+
 #[inline]
 unsafe extern "C" fn pixel_set(dst: *mut pixel, val: libc::c_int, num: libc::c_int) {
     let mut n = 0;
@@ -160,179 +158,6 @@ unsafe extern "C" fn PXSTRIDE(x: ptrdiff_t) -> ptrdiff_t {
         unreachable!();
     }
     return x >> 1;
-}
-#[inline(never)]
-unsafe extern "C" fn padding(
-    mut dst: *mut pixel,
-    mut p: *const pixel,
-    stride: ptrdiff_t,
-    mut left: *const [pixel; 4],
-    mut lpf: *const pixel,
-    mut unit_w: libc::c_int,
-    stripe_h: libc::c_int,
-    edges: LrEdgeFlags,
-) {
-    let have_left =
-        (edges as libc::c_uint & LR_HAVE_LEFT as libc::c_int as libc::c_uint != 0) as libc::c_int;
-    let have_right =
-        (edges as libc::c_uint & LR_HAVE_RIGHT as libc::c_int as libc::c_uint != 0) as libc::c_int;
-    unit_w += 3 * have_left + 3 * have_right;
-    let mut dst_l: *mut pixel = dst.offset((3 * (have_left == 0) as libc::c_int) as isize);
-    p = p.offset(-((3 * have_left) as isize));
-    lpf = lpf.offset(-((3 * have_left) as isize));
-    if edges as libc::c_uint & LR_HAVE_TOP as libc::c_int as libc::c_uint != 0 {
-        let above_1: *const pixel = lpf;
-        let above_2: *const pixel = above_1.offset(PXSTRIDE(stride) as isize);
-        memcpy(
-            dst_l as *mut libc::c_void,
-            above_1 as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_l.offset(390) as *mut libc::c_void,
-            above_1 as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_l.offset((2 * 390) as isize) as *mut libc::c_void,
-            above_2 as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-    } else {
-        memcpy(
-            dst_l as *mut libc::c_void,
-            p as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_l.offset(390) as *mut libc::c_void,
-            p as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_l.offset((2 * 390) as isize) as *mut libc::c_void,
-            p as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        if have_left != 0 {
-            memcpy(
-                dst_l as *mut libc::c_void,
-                &*(*left.offset(0)).as_ptr().offset(1) as *const pixel as *const libc::c_void,
-                ((3 as libc::c_int) << 1) as libc::c_ulong,
-            );
-            memcpy(
-                dst_l.offset(390) as *mut libc::c_void,
-                &*(*left.offset(0)).as_ptr().offset(1) as *const pixel as *const libc::c_void,
-                ((3 as libc::c_int) << 1) as libc::c_ulong,
-            );
-            memcpy(
-                dst_l.offset((2 * 390) as isize) as *mut libc::c_void,
-                &*(*left.offset(0)).as_ptr().offset(1) as *const pixel as *const libc::c_void,
-                ((3 as libc::c_int) << 1) as libc::c_ulong,
-            );
-        }
-    }
-    let mut dst_tl: *mut pixel = dst_l.offset((3 * 390) as isize);
-    if edges as libc::c_uint & LR_HAVE_BOTTOM as libc::c_int as libc::c_uint != 0 {
-        let below_1: *const pixel = lpf.offset(6 * PXSTRIDE(stride));
-        let below_2: *const pixel = below_1.offset(PXSTRIDE(stride) as isize);
-        memcpy(
-            dst_tl.offset((stripe_h * 390) as isize) as *mut libc::c_void,
-            below_1 as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_tl.offset(((stripe_h + 1) * 390) as isize) as *mut libc::c_void,
-            below_2 as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_tl.offset(((stripe_h + 2) * 390) as isize) as *mut libc::c_void,
-            below_2 as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-    } else {
-        let src: *const pixel = p.offset(((stripe_h - 1) as isize * PXSTRIDE(stride)) as isize);
-        memcpy(
-            dst_tl.offset((stripe_h * 390) as isize) as *mut libc::c_void,
-            src as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_tl.offset(((stripe_h + 1) * 390) as isize) as *mut libc::c_void,
-            src as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        memcpy(
-            dst_tl.offset(((stripe_h + 2) * 390) as isize) as *mut libc::c_void,
-            src as *const libc::c_void,
-            (unit_w << 1) as libc::c_ulong,
-        );
-        if have_left != 0 {
-            memcpy(
-                dst_tl.offset((stripe_h * 390) as isize) as *mut libc::c_void,
-                &*(*left.offset((stripe_h - 1) as isize)).as_ptr().offset(1) as *const pixel
-                    as *const libc::c_void,
-                ((3 as libc::c_int) << 1) as libc::c_ulong,
-            );
-            memcpy(
-                dst_tl.offset(((stripe_h + 1) * 390) as isize) as *mut libc::c_void,
-                &*(*left.offset((stripe_h - 1) as isize)).as_ptr().offset(1) as *const pixel
-                    as *const libc::c_void,
-                ((3 as libc::c_int) << 1) as libc::c_ulong,
-            );
-            memcpy(
-                dst_tl.offset(((stripe_h + 2) * 390) as isize) as *mut libc::c_void,
-                &*(*left.offset((stripe_h - 1) as isize)).as_ptr().offset(1) as *const pixel
-                    as *const libc::c_void,
-                ((3 as libc::c_int) << 1) as libc::c_ulong,
-            );
-        }
-    }
-    let mut j = 0;
-    while j < stripe_h {
-        memcpy(
-            dst_tl.offset((3 * have_left) as isize) as *mut libc::c_void,
-            p.offset((3 * have_left) as isize) as *const libc::c_void,
-            (unit_w - 3 * have_left << 1) as libc::c_ulong,
-        );
-        dst_tl = dst_tl.offset(390);
-        p = p.offset(PXSTRIDE(stride) as isize);
-        j += 1;
-    }
-    if have_right == 0 {
-        let mut pad: *mut pixel = dst_l.offset(unit_w as isize);
-        let mut row_last: *mut pixel = &mut *dst_l.offset((unit_w - 1) as isize) as *mut pixel;
-        let mut j_0 = 0;
-        while j_0 < stripe_h + 6 {
-            pixel_set(pad, *row_last as libc::c_int, 3 as libc::c_int);
-            pad = pad.offset(390);
-            row_last = row_last.offset(390);
-            j_0 += 1;
-        }
-    }
-    if have_left == 0 {
-        let mut j_1 = 0;
-        while j_1 < stripe_h + 6 {
-            pixel_set(dst, *dst_l as libc::c_int, 3 as libc::c_int);
-            dst = dst.offset(390);
-            dst_l = dst_l.offset(390);
-            j_1 += 1;
-        }
-    } else {
-        dst = dst.offset((3 * 390) as isize);
-        let mut j_2 = 0;
-        while j_2 < stripe_h {
-            memcpy(
-                dst as *mut libc::c_void,
-                &*(*left.offset(j_2 as isize)).as_ptr().offset(1) as *const pixel
-                    as *const libc::c_void,
-                ((3 as libc::c_int) << 1) as libc::c_ulong,
-            );
-            dst = dst.offset(390);
-            j_2 += 1;
-        }
-    };
 }
 
 unsafe extern "C" fn wiener_c(
@@ -359,7 +184,8 @@ unsafe extern "C" fn wiener_c(
     )
 }
 
-unsafe extern "C" fn wiener_rust(
+// TODO(randompoison): Temporarily public until we can move this to `looprestoration.rs`.
+pub(crate) unsafe extern "C" fn wiener_rust(
     mut p: *mut pixel,
     stride: ptrdiff_t,
     left: *const [pixel; 4],
@@ -372,7 +198,7 @@ unsafe extern "C" fn wiener_rust(
 ) {
     let mut tmp: [pixel; 27300] = [0; 27300];
     let mut tmp_ptr: *mut pixel = tmp.as_mut_ptr();
-    padding(tmp.as_mut_ptr(), p, stride, left, lpf, w, h, edges);
+    padding(&mut tmp, p, stride, left, lpf, w, h, edges);
     let mut hor: [uint16_t; 27300] = [0; 27300];
     let mut hor_ptr: *mut uint16_t = hor.as_mut_ptr();
     let filter: *const [int16_t; 8] = ((*params).filter.0).as_ptr();
@@ -759,7 +585,7 @@ unsafe extern "C" fn sgr_5x5_rust(
 ) {
     let mut tmp: [pixel; 27300] = [0; 27300];
     let mut dst: [coef; 24576] = [0; 24576];
-    padding(tmp.as_mut_ptr(), p, stride, left, lpf, w, h, edges);
+    padding(&mut tmp, p, stride, left, lpf, w, h, edges);
     selfguided_filter(
         dst.as_mut_ptr(),
         tmp.as_mut_ptr(),
@@ -825,7 +651,7 @@ unsafe extern "C" fn sgr_3x3_rust(
 ) {
     let mut tmp: [pixel; 27300] = [0; 27300];
     let mut dst: [coef; 24576] = [0; 24576];
-    padding(tmp.as_mut_ptr(), p, stride, left, lpf, w, h, edges);
+    padding(&mut tmp, p, stride, left, lpf, w, h, edges);
     selfguided_filter(
         dst.as_mut_ptr(),
         tmp.as_mut_ptr(),
@@ -892,7 +718,7 @@ unsafe extern "C" fn sgr_mix_rust(
     let mut tmp: [pixel; 27300] = [0; 27300];
     let mut dst0: [coef; 24576] = [0; 24576];
     let mut dst1: [coef; 24576] = [0; 24576];
-    padding(tmp.as_mut_ptr(), p, stride, left, lpf, w, h, edges);
+    padding(&mut tmp, p, stride, left, lpf, w, h, edges);
     selfguided_filter(
         dst0.as_mut_ptr(),
         tmp.as_mut_ptr(),
